requirements.txt
# ============================================================
#  SIKAI BACKEND — Python Dependencies
#  Install: pip install -r requirements.txt
# ============================================================

# --- Core Framework ---
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# --- AI / LLM ---
groq==0.9.0                    # Groq API (Llama 3.3 70B — ultra fast, cheap)
openai==1.30.0                 # OpenAI-compatible client (also works with Groq)
tiktoken==0.7.0                # Token counting

# --- Database ---
supabase==2.4.6                # Supabase Python client (Postgres + Auth + Storage)
psycopg2-binary==2.9.9         # Direct Postgres driver
sqlalchemy==2.0.30             # ORM for complex queries
alembic==1.13.1                # DB migrations

# --- Cache ---
upstash-redis==1.1.0           # Upstash Redis (serverless, free tier)

# --- Auth & Security ---
python-jose[cryptography]==3.3.0   # JWT tokens
passlib[bcrypt]==1.7.4             # Password hashing
python-dotenv==1.0.1               # .env loading

# --- Validation ---
pydantic==2.7.1                # Data validation (FastAPI uses this)
pydantic-settings==2.2.1       # Settings management

# --- HTTP / Async ---
httpx==0.27.0                  # Async HTTP client
aiofiles==23.2.1               # Async file I/O

# --- Utilities ---
loguru==0.7.2                  # Beautiful structured logging
python-slugify==8.0.4          # Slug generation for course URLs
Pillow==10.3.0                 # Image processing (for thumbnails)

# --- Rate Limiting ---
slowapi==0.1.9                 # Rate limiting for FastAPI

# --- Testing ---
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0                  # Also used for testing
